lead_optimization/requirements.txt
mmpdb==3.1.3
rdkit==2025.3.5
psycopg[binary]>=3.2.0
pyyaml>=6.0
requests>=2.28.0
